Marconi.ChainIndex.CLI

Documentation

chainPointParser :: Parser ChainPoint

fromJustWithError :: Show e => Either e a -> a

Exit program with error Note, if the targetAddress parser fails, or is empty, there is nothing to do for the hotStore. In such case we should fail fast

pNetworkId :: Parser NetworkId

pMainnet :: Parser NetworkId

pTestnetMagic :: Parser NetworkMagic

multiString :: Mod OptionFields [Address ShelleyAddr] -> Parser TargetAddresses

parses CLI params to valid NonEmpty list of Shelley addresses We error out if there are any invalid addresses

parseCardanoAddresses :: String -> [Address ShelleyAddr]

data Options

This executable is meant to exercise a set of indexers (for now datumhash -> datum) against the mainnet (meant to be used for testing).

In case you want to access the results of the datumhash indexer you need to query the resulting database: $ sqlite3 datums.sqlite > select slotNo, datumHash, datum from kv_datumhsh_datum where slotNo = 39920450; 39920450|679a55b523ff8d61942b2583b76e5d49498468164802ef1ebe513c685d6fb5c2|X(002f9787436835852ea78d3c45fc3d436b324184

Constructors

Options

Fields

optionsSocketPath :: !String

optionsNetworkId :: !NetworkId

optionsChainPoint :: !ChainPoint

optionsDbPath :: !FilePath

SQLite database directory path

optionsDisableUtxo :: !Bool

optionsDisableAddressDatum :: !Bool

optionsDisableDatum :: !Bool

optionsDisableScript :: !Bool

optionsDisableEpochState :: !Bool

optionsDisableMintBurn :: !Bool

optionsTargetAddresses :: !(Maybe TargetAddresses)

optionsNodeConfigPath :: !(Maybe FilePath)

parseOptions :: IO Options

programParser :: String -> ParserInfo Options

optionsParser :: Parser Options

optAddressesParser :: Mod OptionFields [Address ShelleyAddr] -> Parser (Maybe TargetAddresses)

Database paths

utxoDbPath :: Options -> Maybe FilePath

addressDatumDbPath :: Options -> Maybe FilePath

datumDbPath :: Options -> Maybe FilePath

scriptTxDbPath :: Options -> Maybe FilePath

epochStateDbPath :: Options -> Maybe FilePath

mintBurnDbPath :: Options -> Maybe FilePath

Common CLI parsers for other derived programs.

commonSocketPath :: Parser String

commonDbDir :: Parser String

commonVersionOption :: String -> Parser (a -> a)

getGitSha :: IO String

marconiDescr :: String -> InfoMod a

commonMaybePort :: Parser (Maybe Int)

commonMaybeTargetAddress :: Parser (Maybe TargetAddresses)