Marconi.ChainIndex.Indexers

Documentation

data AnyTxBody

Constructors

forall era.IsCardanoEra era => AnyTxBody BlockNo TxIndexInBlock (TxBody era)

type Coordinator = WithTrace IO Coordinator TipOrBlock

type ChainTipIndexer = ChainTipIndexer IO

type EpochStateIndexer = EpochStateIndexer (WithDistance (Maybe ExtLedgerState, BlockInMode CardanoMode))

type MintTokenEventIndexer = StandardIndexer IO SQLiteIndexer MintTokenBlockEvents

type BlockInfoIndexer = StandardIndexer IO SQLiteIndexer BlockInfo

type UtxoIndexer = UtxoQueryIndexer IO

type CurrentSyncPointIndexer = CurrentSyncPointQueryIndexer TipOrBlock

data MarconiChainIndexQueryables

Container for all the queryable indexers of marconi-chain-index.

Constructors

MarconiChainIndexQueryables

Fields

_queryableEpochState :: !(MVar EpochStateIndexer)

_queryableMintToken :: !(MVar MintTokenEventIndexer)

_queryableUtxo :: !UtxoIndexer

_queryableCurrentSyncPoint :: !CurrentSyncPointIndexer

queryableUtxo :: Lens' MarconiChainIndexQueryables UtxoIndexer

queryableMintToken :: Lens' MarconiChainIndexQueryables (MVar MintTokenEventIndexer)

queryableEpochState :: Lens' MarconiChainIndexQueryables (MVar EpochStateIndexer)

queryableCurrentSyncPoint :: Lens' MarconiChainIndexQueryables CurrentSyncPointIndexer

buildIndexers :: SecurityParam -> CatchupConfig -> UtxoIndexerConfig -> MintTokenEventConfig -> EpochStateWorkerConfig -> Trace IO Text -> FilePath -> ExceptT IndexerError IO (ChainPoint, MarconiChainIndexQueryables, Coordinator)

Build all the indexers of marconi-chain-index (all those which are available with the new implementation) and expose a single coordinator to operate them

buildBlockEventCoordinator :: MonadIO m => Trace IO (IndexerEvent ChainPoint) -> [Worker (WithDistance BlockEvent) ChainPoint] -> m (Worker TipOrBlock ChainPoint)

Build and start a coordinator of a bunch of workers that takes an AnyTxBody as an input

buildTxBodyCoordinator :: (MonadIO m, Ord (Point event), Show (Point event)) => Trace IO Text -> (WithDistance input -> IO (Maybe event)) -> [Worker event (Point event)] -> m (Worker (WithDistance input) (Point event))

Build and start a coordinator of a bunch of workers that takes an AnyTxBody as an input

blockInfoBuilder :: (MonadIO n, MonadError IndexerError n, MonadIO m) => SecurityParam -> CatchupConfig -> Trace m Text -> FilePath -> n (StandardWorker m BlockEvent BlockInfo SQLiteIndexer)

Configure and start the BlockInfo indexer

utxoBuilder :: (MonadIO n, MonadError IndexerError n) => SecurityParam -> CatchupConfig -> UtxoIndexerConfig -> Trace IO Text -> FilePath -> n (StandardWorker IO [AnyTxBody] UtxoEvent SQLiteIndexer)

Configure and start the Utxo indexer

chainTipBuilder :: (MonadIO n, MonadError IndexerError n, MonadIO m) => Trace m (IndexerEvent ChainPoint) -> FilePath -> n (WorkerIndexer m TipOrBlock ChainTip (WithTrace m LastEventIndexer))

Configure and start the ChainTip indexer

spentBuilder :: (MonadIO n, MonadError IndexerError n) => SecurityParam -> CatchupConfig -> Trace IO Text -> FilePath -> n (StandardWorker IO [AnyTxBody] SpentInfoEvent SQLiteIndexer)

Configure and start the SpentInfo indexer

datumBuilder :: (MonadIO n, MonadError IndexerError n, MonadIO m) => SecurityParam -> CatchupConfig -> Trace m Text -> FilePath -> n (StandardWorker m [AnyTxBody] DatumEvent SQLiteIndexer)

Configure and start the Datum indexer

mintBuilder :: (MonadIO n, MonadError IndexerError n) => SecurityParam -> CatchupConfig -> MintTokenEventConfig -> Trace IO Text -> FilePath -> n (StandardWorker IO [AnyTxBody] MintTokenBlockEvents SQLiteIndexer)

Configure and start the MintToken indexer

epochStateBuilder :: (MonadIO n, MonadError IndexerError n) => SecurityParam -> CatchupConfig -> EpochStateWorkerConfig -> Trace IO Text -> FilePath -> n (WorkerIndexer IO (WithDistance BlockEvent) (WithDistance (Maybe ExtLedgerState, BlockInMode CardanoMode)) EpochStateIndexer)

Configure and start the EpochState indexer