Marconi.ChainIndex.Indexers.AddressDatum

Description

Module for indexing the datums for all addresses in the Cardano blockchain.

AddressDatumIndex

type AddressDatumIndex = State AddressDatumHandle

data AddressDatumHandle

Define the handler data type, meant as a wrapper for the connection type (in this case the SQLite connection). In this indexer, we also add the number of events that we want to return from the on-disk buffer.

data family StorableEvent h

data family StorableQuery h

data family StorableResult h

toAddressDatumIndexEvent :: Maybe (Address ShelleyAddr -> Bool) -> [Tx era] -> ChainPoint -> StorableEvent AddressDatumHandle

type AddressDatumQuery = StorableQuery AddressDatumHandle

type AddressDatumResult = StorableResult AddressDatumHandle

newtype AddressDatumDepth

Constructors

AddressDatumDepth Int

open :: FilePath -> AddressDatumDepth -> IO AddressDatumIndex