Marconi.ChainIndex.Indexers.EpochState

Description

Module for indexing the stakepool delegation per epoch in the Cardano blockchain.

This module will create the SQL tables:

+ table: epoch_sdd

   |---------+--------+----------+--------+-----------------+---------|
   | epochNo | poolId | lovelace | slotNo | blockHeaderHash | blockNo |
   |---------+--------+----------+--------+-----------------+---------|

+ table: epoch_nonce

   |---------+-------+--------+-----------------+---------|
   | epochNo | nonce | slotNo | blockHeaderHash | blockNo |
   |---------+-------+--------+-----------------+---------|

To create those tables, we need to compute the ExtLedgerState from `ouroboros-network` (combination of `LedgerState, called NewEpochState in `cardano-ledger`, and the HeaderState) at each Rollforward chain sync event. Using the ExtLegderState, we can easily compute the epoch nonce as well as the stake pool delegation for that epoch.

The main issue with this indexer is that building the ExtLedgerState and saving it on disk for being able to resume is VERY resource intensive. Syncing time for this indexer is over 20h and uses about ~16GB of RAM (which will keep increasing as the blockchain continues to grow).

Here is a synopsis of what this indexer does.

We assume that the construction of LedgerState is done outside of this indexer (this module).

the insert function is called with the *last* event of an epoch (therefore, the last LedgerState before starting a new epoch). We do that because we only care about the SDD (Stake Pool Delegation) from the last block before a new epoch.

Once the StorableEvent is stored on disk, we perform various steps:

we save the SDD for the current epoch in the epoch_sdd table

we save the LedgerStates in the filesystem as binary files (the ledger state file path has the format: `ledgerState_SLOT_NO_BLOCK_HEADER_HASH_BLOCK_NO.bin`). We only store a LedgerState if it's rollbackable or if the last one of a given epoch. This step is necessary for resuming the indexer.

we delete immutable LedgerState binary files expect latest one (this step is necessary for

The indexer provides the following queries:

C.EpochNo -> SDD (the actualy query that clients will be interested in)

C.ChainPoint -> LedgerState (query that is necessary for resuming)

EpochStateIndex

type EpochStateIndex = State EpochStateHandle

data EpochStateHandle

data EpochSDDRow

Constructors

EpochSDDRow

Fields

epochSDDRowEpochNo :: !EpochNo

epochSDDRowPoolId :: !PoolId

epochSDDRowLovelace :: !Lovelace

epochSDDRowSlotNo :: !SlotNo

epochSDDRowBlockHeaderHash :: !(Hash BlockHeader)

epochSDDRowBlockNo :: !BlockNo

data EpochNonceRow

Constructors

EpochNonceRow

Fields

epochNonceRowEpochNo :: !EpochNo

epochNonceRowNonce :: !Nonce

epochNonceRowSlotNo :: !SlotNo

epochNonceRowBlockHeaderHash :: !(Hash BlockHeader)

epochNonceRowBlockNo :: !BlockNo

data family StorableEvent h

data family StorableQuery h

data family StorableResult h

toStorableEvent

Arguments

:: ExtLedgerState (HardForkBlock (CardanoEras StandardCrypto))

-> SlotNo

-> Hash BlockHeader

-> BlockNo

-> ChainTip

-> SecurityParam

-> Bool
Is the last event of the current epoch

-> StorableEvent EpochStateHandle

open

Arguments

:: TopLevelConfig (CardanoBlock StandardCrypto)

-> FilePath
SQLite database file path

-> FilePath
Directory from which we will save the various LedgerState as different points in time.

-> SecurityParam

-> IO (State EpochStateHandle)

getEpochNo :: ExtLedgerState (CardanoBlock StandardCrypto) -> Maybe EpochNo