Marconi.ChainIndex.Indexers.MintTokenEvent

Description

Mint/burn event indexer, the result of which is an sqlite database 'minttokenevent.db' which has a table minting_policy_events with the following fields:

( slotNo INT NOT NULL
, blockHeaderHash BLOB NOT NULL
, blockNo INT NOT NULL
, txId BLOB NOT NULL
, txIndexInBlock INT NOT NULL
, policyId BLOB NOT NULL
, assetName TEXT NOT NULL
, quantity INT NOT NULL
, redeemerHash BLOB
, redeemerData BLOB
)

Events

newtype MintTokenBlockEvents

Minting events given for each block.

Constructors

MintTokenBlockEvents (NonEmpty MintTokenEvent)

mintTokenEvents :: Iso' MintTokenBlockEvents (NonEmpty MintTokenEvent)

data MintTokenEvent

Single minting event. This is the datatype was will be used to store the events in the database (not MintTokenBlockEvents). More specifically, we will store 'Core.Timed (Core.Point MintTokenEvent) MintTokenEvent)'.

Constructors

MintTokenEvent !MintTokenEventLocation !MintAsset

mintTokenEventLocation :: Lens' MintTokenEvent MintTokenEventLocation

mintTokenEventAsset :: Lens' MintTokenEvent MintAsset

data MintTokenEventLocation

Constructors

MintTokenEventLocation !BlockNo !TxId !TxIndexInBlock

mintTokenEventBlockNo :: Lens' MintTokenEventLocation BlockNo

mintTokenEventIndexInBlock :: Lens' MintTokenEventLocation TxIndexInBlock

mintTokenEventTxId :: Lens' MintTokenEventLocation TxId

data MintAsset

Constructors

MintAsset !PolicyId !AssetName !Quantity !(Maybe MintAssetRedeemer)

mintAssetAssetId :: Lens' MintAsset AssetId

mintAssetPolicyId :: Lens' MintAsset PolicyId

mintAssetAssetName :: Lens' MintAsset AssetName

mintAssetQuantity :: Lens' MintAsset Quantity

mintAssetRedeemer :: Lens' MintAsset (Maybe MintAssetRedeemer)

data MintAssetRedeemer

Constructors

MintAssetRedeemer !ScriptData !(Hash ScriptData)

mintAssetRedeemerData :: Lens' MintAssetRedeemer ScriptData

mintAssetRedeemerHash :: Lens' MintAssetRedeemer (Hash ScriptData)

Indexer and worker

type MintTokenEventIndexer = SQLiteIndexer MintTokenBlockEvents

A raw SQLite indexer for MintTokenBlockEvents

type StandardMintTokenEventIndexer m = StandardSQLiteIndexer m MintTokenBlockEvents

A SQLite MintTokenBlockEvents indexer with Catchup

newtype MintTokenEventConfig

MintTokenEventConfig allows for specifying a list of C.PolicyIds and possibly C.AssetNames by which to filter a query. Nothing represents the case in which no filtering should occur.

Constructors

MintTokenEventConfig

Fields

_configTrackedAssetIds :: Maybe (NonEmpty (PolicyId, Maybe AssetName))

configTrackedAssetIds :: Iso' MintTokenEventConfig (Maybe (NonEmpty (PolicyId, Maybe AssetName)))

catchupConfigEventHook :: Text -> Trace IO Text -> FilePath -> CatchupEvent -> IO ()

mkMintTokenIndexer :: (MonadIO m, MonadError IndexerError m) => FilePath -> m MintTokenEventIndexer

mkMintTokenEventWorker

Arguments

:: (MonadIO n, MonadError IndexerError n, MonadIO m)

=> StandardWorkerConfig m input MintTokenBlockEvents
General configuration of the indexer (mostly for logging purpose)

-> MintTokenEventConfig
Specific configuration of the indexer (mostly for logging purpose and filtering for target asset ids)

-> FilePath

-> n (StandardWorker m input MintTokenBlockEvents SQLiteIndexer)

Create a worker for the MintTokenEvent indexer

filterByTargetAssetIds :: NonEmpty (PolicyId, Maybe AssetName) -> MintTokenBlockEvents -> Maybe MintTokenBlockEvents

Only keep the MintTokenEvents at a block if they mint a target AssetId.

Extract events

extractEventsFromTx :: BlockNo -> TxIndexInBlock -> TxBody era -> [MintTokenEvent]

Queries

data MintTokenEventsMatchingQuery event

Defunctionalisation of queries

Constructors

AllEvents

AllMintEvents

AllBurnEvents

ByTargetAssetIds ByTargetAssetIdsArgs

newtype ByTargetAssetIdsArgs

Constructors

ByTargetAssetIdsArgs (NonEmpty (PolicyId, Maybe AssetName))

evalMintTokenEventsMatchingQuery :: MintTokenEventsMatchingQuery MintTokenBlockEvents -> MintTokenBlockEvents -> Maybe MintTokenBlockEvents

Evaluator for MintTokenEventsMatchingQuery

data QueryByAssetId event

Constructors

QueryByAssetId

Fields

_queryByAssetIdPolicyId :: !PolicyId

_queryByAssetIdAssetName :: !(Maybe AssetName)

_queryByAssetIdEventType :: !(Maybe EventType)

_queryByAssetIdUpperSlotNo :: !(Maybe SlotNo)

_queryByAssetIdLowerTxId :: !(Maybe TxId)

data EventType

Constructors

MintEventType

BurnEventType

toTimedMintEvents :: [Timed ChainPoint MintTokenEvent] -> [Timed ChainPoint MintTokenBlockEvents]