Marconi.ChainIndex.Indexers.Utxo

Description

Module for indexing the Utxos in the Cardano blockchain

Documentation

type UtxoIndexer = State UtxoHandle

data UtxoHandle

Constructors

UtxoHandle

Fields

hdlConnection :: !Connection

SQLite connection

hdlDpeth :: !Int

depth before flushing to disk storage

toVacuume :: !Bool

weather to perform SQLite vacuum to release space

type QueryableAddresses = NonEmpty (StorableQuery UtxoHandle)

newtype Depth

Constructors

Depth Int

data Utxo

Constructors

Utxo

Fields

_address :: !AddressAny

_txId :: !TxId

_txIx :: !TxIx

_datum :: !(Maybe ScriptData)

_datumHash :: !(Maybe (Hash ScriptData))

_value :: !Value

_inlineScript :: !(Maybe ScriptInAnyLang)

_inlineScriptHash :: !(Maybe ScriptHash)

value :: Lens' Utxo Value

txIx :: Lens' Utxo TxIx

txId :: Lens' Utxo TxId

inlineScriptHash :: Lens' Utxo (Maybe ScriptHash)

inlineScript :: Lens' Utxo (Maybe ScriptInAnyLang)

datumHash :: Lens' Utxo (Maybe (Hash ScriptData))

datum :: Lens' Utxo (Maybe ScriptData)

address :: Lens' Utxo AddressAny

newtype ChainPointRow

Constructors

ChainPointRow

Fields

getChainPoint :: ChainPoint

data UtxoRow

Constructors

UtxoRow

Fields

_urUtxo :: !Utxo

_urSlotNo :: !SlotNo

_urBlockHash :: !(Hash BlockHeader)

urUtxo :: Lens' UtxoRow Utxo

urSlotNo :: Lens' UtxoRow SlotNo

urBlockHash :: Lens' UtxoRow (Hash BlockHeader)

eventIsBefore :: ChainPoint -> StorableEvent UtxoHandle -> Bool

data TxOutBalance

The effect of a transaction (or a number of them) on the tx output set.

Constructors

TxOutBalance

Fields

_tobUnspent :: !(Set TxIn)

Outputs newly added by the transaction(s)

_tobSpent :: !(Set TxIn)

Outputs spent by the transaction(s)

data Spent

Constructors

Spent

Fields

_sTxId :: !TxId

_sTxIx :: !TxIx

_sSlotNo :: !SlotNo

_sBlockHash :: !(Hash BlockHeader)

sTxIx :: Lens' Spent TxIx

sTxId :: Lens' Spent TxId

sSlotNo :: Lens' Spent SlotNo

sBlockHash :: Lens' Spent (Hash BlockHeader)

open

Arguments

:: FilePath
sqlite file path

-> Depth
The Depth parameter k, the larger K, the more RAM the indexer uses

-> Bool
whether to perform vacuum

-> IO UtxoIndexer

Open a connection to DB, and create resources The parameter ((k + 1) * 2) specifies the amount of events that are buffered. The larger the number, the more RAM the indexer uses. However, we get improved SQL queries due to batching more events together.

getSpentFrom :: StorableEvent UtxoHandle -> [Spent]

getTxIns :: Connection -> SlotNo -> IO (Set TxIn)

Retrieve TxIns at a slotNo This function is used to reconstruct the original UtxoEvent

rowsToEvents

Arguments

:: (SlotNo -> IO (Set TxIn))
function to fetch TxIn

-> [UtxoRow]
rows to convert back to event

-> IO [StorableEvent UtxoHandle]
utxo events

Convert UtxoRows to UtxoEvents

mergeInMemoryAndSql :: Foldable f => f (StorableEvent UtxoHandle) -> [UtxoRow] -> [UtxoRow]

merge in-memory events with SQL retreived UtxoRows Notes, a peroperty of this merge is to remove all spent utxos from the resulting [UtxoRow]

eventToRows :: StorableEvent UtxoHandle -> [UtxoRow]

convert utxoEvent to utxoRow Note: No unspent computeation is performed

eventsAtAddress

Arguments

:: Foldable f

=> AddressAny
Address query

-> Maybe SlotNo
Latest included chainpoint

-> f (StorableEvent UtxoHandle)
Utxo event

-> [StorableEvent UtxoHandle]
Utxo event at thegiven address

Filter for events at the given address

addressFilteredRows

Arguments

:: Foldable f

=> AddressAny
query

-> Maybe SlotNo
latest included chainpoint

-> f (StorableEvent UtxoHandle)
Utxo Event

-> [UtxoRow]
Rows at the query

only store rows in the address list.

utxoAtAddressQuery

Arguments

:: Foldable f

=> Connection

-> f (StorableEvent UtxoHandle)

-> StorableEvent UtxoHandle

-> [Query]
the filter part of the query

-> [NamedParam]

-> IO (StorableResult UtxoHandle)

queryBuffer

Arguments

:: Foldable f

=> QueryInterval ChainPoint

-> AddressAny
Query

-> Maybe SlotNo
Latest included point

-> f (StorableEvent UtxoHandle)
Utxo events

-> StorableEvent UtxoHandle

Query memory buffer

toAddr :: AddressInEra era -> AddressAny

Convert from AddressInEra of the CurrentEra to AddressAny.

getUtxoEventsFromBlock

Arguments

:: IsCardanoEra era

=> Maybe TargetAddresses
target addresses to filter for

-> Block era

-> StorableEvent UtxoHandle
UtxoEvents are stored in storage after conversion to UtxoRow

Extract UtxoEvents from Cardano Block

getUtxoEvents

Arguments

:: IsCardanoEra era

=> Maybe TargetAddresses
target addresses to filter for

-> [Tx era]

-> ChainPoint

-> StorableEvent UtxoHandle
UtxoEvents are stored in storage after conversion to UtxoRow

Extract UtxoEvents from Cardano Transactions

isAddressInTarget :: Maybe TargetAddresses -> AddressAny -> Bool

does the transaction contain a targetAddress

getUtxos :: IsCardanoEra era => Maybe TargetAddresses -> Tx era -> Map TxIn Utxo

getRefScriptAndHash :: ReferenceScript era -> (Maybe ScriptInAnyLang, Maybe ScriptHash)

get the inlineScript and inlineScriptHash

getScriptDataAndHash :: TxOutDatum CtxTx era -> (Maybe ScriptData, Maybe (Hash ScriptData))

Get the datum hash and datum or a transaction output.

rmSpent :: Set TxIn -> [Utxo] -> [Utxo]

remove spent transactions

getInputs :: Tx era -> Set TxIn

txScriptValidityToScriptValidity :: TxScriptValidity era -> ScriptValidity

Duplicated from cardano-api (not exposed in cardano-api) This function should be removed when marconi will depend on a cardano-api version that has accepted this PR: https://github.com/input-output-hk/cardano-node/pull/4569

isAddressInTarget' :: TargetAddresses -> Utxo -> Bool

does the transaction contain a targetAddress

mkQueryableAddresses :: TargetAddresses -> QueryableAddresses

txOutBalanceFromTxs :: [Tx era] -> TxOutBalance

txOutBalanceFromTx :: Tx era -> TxOutBalance

convertTxOutToUtxo :: TxId -> TxIx -> TxOut CtxTx BabbageEra -> Utxo