Marconi.ChainIndex.Indexers.UtxoQuery

Description

An aggregate indexer that gathers information of Utxo, Spent, Datum and BlockInfo to provide the UTxO information with resolved datum and information about the TxIn used to produce an UTxO.

Indexer

type UtxoQueryIndexer m = SQLiteAggregateQuery m ChainPoint UtxoQueryEvent

An alias for the SQLiteAggregateQuery that handle the UtxoQueryEvent

data UtxoQueryAggregate m

Explicit list of resources needed for the UtxoQueryAggregate

Constructors

forall utxoIndexer utxo spentIndexer spent datumIndexer datum blockInfoIndexer blockInfo.(IsSourceProvider m (NonEmpty Utxo) utxoIndexer, IsSourceProvider m utxo utxoIndexer, Point utxo ~ Point (NonEmpty Utxo), IsSourceProvider m (NonEmpty SpentInfo) spentIndexer, IsSourceProvider m spent spentIndexer, Point spent ~ Point (NonEmpty SpentInfo), IsSourceProvider m (NonEmpty DatumInfo) datumIndexer, IsSourceProvider m datum datumIndexer, Point datum ~ Point (NonEmpty DatumInfo), IsSourceProvider m BlockInfo blockInfoIndexer, IsSourceProvider m blockInfo blockInfoIndexer, Point blockInfo ~ Point BlockInfo) => UtxoQueryAggregate

Fields

utxoIndexer :: MVar (utxoIndexer utxo)

the source provider for the Utxo table, usually a UtxoIndexer

spentIndexer :: MVar (spentIndexer spent)

the source provider for the Spent table, usually a SpentIndexer

datumIndexer :: MVar (datumIndexer datum)

the source provider for the Datum table, usually a DatumIndexer

blockInfoIndexer :: MVar (blockInfoIndexer blockInfo)

the source provider for the BlockInfo table, usually a BlockInfoIndexer

data UtxoQueryEvent

Uninhabited type for the UtxoQueryAggregate (as it's an aggregate, it doesn't index anything)

mkUtxoSQLiteQuery :: UtxoQueryAggregate m -> IO (UtxoQueryIndexer m)

Generate a UtxoQueryIndexer from the given source

Query

data UtxoQueryInput

Constructors

UtxoQueryInput

Fields

_address :: AddressAny

_lowerBound :: Maybe SlotNo

Inclusive lowerBound

_upperBound :: Maybe SlotNo

Inclusive upperBound (utxo spent after this point displaed with their spent information)

address :: Lens' UtxoQueryInput AddressAny

lowerBound :: Lens' UtxoQueryInput (Maybe SlotNo)

upperBound :: Lens' UtxoQueryInput (Maybe SlotNo)

Result

data UtxoResult

Constructors

UtxoResult Utxo !(Maybe ScriptData) !(Timed ChainPoint BlockInfo) !(Maybe (Timed ChainPoint (BlockInfo, TxId))) [TxIn]

utxo :: Lens' UtxoResult Utxo

datum :: Lens' UtxoResult (Maybe ScriptData)

blockInfo :: Lens' UtxoResult (Timed ChainPoint BlockInfo)

spentInfo :: Lens' UtxoResult (Maybe (Timed ChainPoint (BlockInfo, TxId)))

inputs :: Lens' UtxoResult [TxIn]