Marconi.ChainIndex.Indexers.Worker

Description

Common worker configuration and creation

Documentation

type StandardIndexer m indexer event = WithTrace m (WithCatchup (WithTransform indexer event)) (WithDistance (Maybe event))

An alias for an indexer with catchup and transformation to perform filtering

type StandardSQLiteIndexer m event = StandardIndexer m SQLiteIndexer event

An alias for an SQLiteWorker with catchup and transformation to perform filtering

data StandardWorkerConfig m input event

Constructors

StandardWorkerConfig

Fields

workerName :: Text

securityParamConfig :: !SecurityParam

catchupConfig :: CatchupConfig

eventExtractor :: input -> m (Maybe event)

logger :: Trace m (IndexerEvent ChainPoint)

data StandardWorker m input event indexer

Contains a worker for a given indexer and the MVar that is modified by this worker

Constructors

StandardWorker

Fields

standardWorkerIndexerVar :: !(MVar (StandardIndexer m indexer event))

standardWorker :: WorkerM m (WithDistance input) (Point event)

mkStandardWorker :: (MonadIO m, MonadIO n, WorkerIndexerType (ExceptT IndexerError m) event indexer, IsSync n event indexer, Point event ~ ChainPoint) => StandardWorkerConfig m input event -> indexer event -> n (StandardWorker m input event indexer)

Create a worker for the given indexer with some standard catchup values

mkStandardWorkerWithFilter :: (MonadIO n, MonadIO m, WorkerIndexerType (ExceptT IndexerError m) event indexer, IsSync n event indexer, Ord (Point event), Point event ~ ChainPoint) => StandardWorkerConfig m input event -> (event -> Maybe event) -> indexer event -> n (StandardWorker m input event indexer)

Create a worker for the given indexer with some standard catchup values with extra filtering.

mkStandardIndexer :: (MonadIO m, Point event ~ ChainPoint) => StandardWorkerConfig m a b -> indexer event -> StandardIndexer m indexer event

Create a standard indexer from the given indexer with some standard catchup values.

mkStandardIndexerWithFilter :: (MonadIO m, Point event ~ ChainPoint) => StandardWorkerConfig m a b -> (event -> Maybe event) -> indexer event -> StandardIndexer m indexer event

Create a standard indexer from the given indexer with some standard catchup values with extra filtering.