Marconi.ChainIndex.Logger

Description

Common loggers for marconi

Documentation

nullTracer :: forall (m :: Type -> Type) a. Applicative m => Tracer m a

defaultStdOutLogger :: Text -> IO (Trace IO Text, Switchboard Text)

StdOut logger, only log stuff above the Info level

mkMarconiTrace :: Trace m Text -> MarconiTrace m

Builds a MarconiTrace from a base tracer.

chainSyncEventStreamLogging :: MarconiTrace IO -> Stream (Of (ChainSyncEvent BlockEvent)) IO r -> Stream (Of (ChainSyncEvent BlockEvent)) IO r

data LastSyncLog

Logging datatype for information that occured since the previous LastSyncLog.

Constructors

LastSyncLog

Fields

syncStatsSyncLog :: LastSyncStats

Stats since the last syncing log message

timeSinceLastMsgSyncLog :: Maybe NominalDiffTime

Time since last syncing log message.

data LastSyncStats

Chain synchronisation statistics measured starting from previously measured LastSyncStats.

Constructors

LastSyncStats

Fields

syncStatsNumBlocks :: !Word64

Number of applied blocks since last message

syncStatsNumRollbacks :: !Word64

Number of rollbacks since last message

syncStatsChainSyncPoint :: ChainPoint

Chain index syncing point

syncStatsNodeTip :: ChainTip

Current node tip

syncStatsLastMessageTime :: !(Maybe UTCTime)

Timestamp of last printed message

marconiFormatting :: Doc ann -> Text