Marconi.ChainIndex.Node.Client.GenesisConfig

Description

This module is a duplication of functionality from the LedgerState module in the `cardano-api` Haskell project which is not exposed .

TODO: If `cardano-api` ever exposes this module, then we should use that.

Documentation

data GenesisConfig

Constructors

GenesisCardano !NodeConfig !Config !ShelleyConfig !AlonzoGenesis

data ShelleyGenesisError

Constructors

ShelleyGenesisReadError !FilePath !Text

ShelleyGenesisHashMismatch !GenesisHashShelley !GenesisHashShelley

ShelleyGenesisDecodeError !FilePath !Text

data AlonzoGenesisError

Constructors

AlonzoGenesisReadError !FilePath !Text

AlonzoGenesisHashMismatch !GenesisHashAlonzo !GenesisHashAlonzo

AlonzoGenesisDecodeError !FilePath !Text

renderAlonzoGenesisError :: AlonzoGenesisError -> Text

renderShelleyGenesisError :: ShelleyGenesisError -> Text

renderHash :: Hash Blake2b_256 ByteString -> Text

data ShelleyConfig

Constructors

ShelleyConfig

Fields

scConfig :: !(ShelleyGenesis StandardShelley)

scGenesisHash :: !GenesisHashShelley

newtype GenesisFile

Constructors

GenesisFile

Fields

unGenesisFile :: FilePath

newtype GenesisHashByron

Constructors

GenesisHashByron

Fields

unGenesisHashByron :: Text

newtype GenesisHashShelley

Constructors

GenesisHashShelley

Fields

unGenesisHashShelley :: Hash Blake2b_256 ByteString

newtype GenesisHashAlonzo

Constructors

GenesisHashAlonzo

Fields

unGenesisHashAlonzo :: Hash Blake2b_256 ByteString

newtype LedgerStateDir

Constructors

LedgerStateDir

Fields

unLedgerStateDir :: FilePath

newtype NetworkName

Constructors

NetworkName

Fields

unNetworkName :: Text

newtype NetworkConfigFile

Constructors

NetworkConfigFile

Fields

unNetworkConfigFile :: FilePath

newtype SocketPath

Constructors

SocketPath

Fields

unSocketPath :: FilePath

data NodeConfig

Constructors

NodeConfig

Fields

ncPBftSignatureThreshold :: !(Maybe Double)

ncByronGenesisFile :: !GenesisFile

ncByronGenesisHash :: !GenesisHashByron

ncShelleyGenesisFile :: !GenesisFile

ncShelleyGenesisHash :: !GenesisHashShelley

ncAlonzoGenesisFile :: !GenesisFile

ncAlonzoGenesisHash :: !GenesisHashAlonzo

ncRequiresNetworkMagic :: !RequiresNetworkMagic

ncByronSoftwareVersion :: !SoftwareVersion

ncByronProtocolVersion :: !ProtocolVersion

ncByronToShelley :: !(ProtocolTransitionParamsShelleyBased StandardShelley)

ncShelleyToAllegra :: !(ProtocolTransitionParamsShelleyBased StandardAllegra)

ncAllegraToMary :: !(ProtocolTransitionParamsShelleyBased StandardMary)

ncMaryToAlonzo :: !TriggerHardFork

ncAlonzoToBabbage :: !TriggerHardFork

readNetworkConfig :: NetworkConfigFile -> ExceptT Text IO NodeConfig

adjustGenesisFilePath :: (FilePath -> FilePath) -> GenesisFile -> GenesisFile

mkAdjustPath :: FilePath -> FilePath -> FilePath

parseNodeConfig :: ByteString -> Either Text NodeConfig

readCardanoGenesisConfig :: NodeConfig -> ExceptT GenesisConfigError IO GenesisConfig

readByronGenesisConfig :: NodeConfig -> ExceptT GenesisConfigError IO Config

readShelleyGenesisConfig :: NodeConfig -> ExceptT GenesisConfigError IO ShelleyConfig

readAlonzoGenesisConfig :: NodeConfig -> ExceptT GenesisConfigError IO AlonzoGenesis

readAlonzoGenesis :: GenesisFile -> GenesisHashAlonzo -> ExceptT AlonzoGenesisError IO AlonzoGenesis

readShelleyGenesis :: GenesisFile -> GenesisHashShelley -> ExceptT ShelleyGenesisError IO ShelleyConfig

data GenesisConfigError

Constructors

NEError !Text

NEByronConfig !FilePath !ConfigurationError

NEShelleyConfig !FilePath !Text

NEAlonzoConfig !FilePath !Text

NECardanoConfig !Text

renderGenesisConfigError :: GenesisConfigError -> Text

initExtLedgerStateVar :: GenesisConfig -> ExtLedgerState (HardForkBlock (CardanoEras StandardCrypto))

mkProtocolInfoCardano :: GenesisConfig -> ProtocolInfo IO (HardForkBlock (CardanoEras StandardCrypto))

shelleyProtVer :: NodeConfig -> ProtVer

shelleyPraosNonce :: ShelleyConfig -> Nonce

textShow :: Show a => a -> Text

readByteString :: FilePath -> Text -> ExceptT Text IO ByteString