Marconi.ChainIndex.Transformer.WithSyncLog

Documentation

newtype WithSyncStats indexer event

A logging modifier that adds stats logging to the indexer

Constructors

WithSyncStats

Fields

_syncStatsWrapper :: IndexTransformer SyncLog indexer event

data LastSyncStats

Chain synchronisation statistics measured starting from previously measured LastSyncStats.

Constructors

LastSyncStats

Fields

_syncStatsNumBlocks :: !Word64

Number of applied blocks since last message

_syncStatsNumRollbacks :: !Word64

Number of rollbacks since last message

_syncStatsChainSyncPoint :: ChainPoint

Chain index syncing point

_syncStatsNodeTip :: ChainTip

Current node tip

_syncStatsLastMessageTime :: !(Maybe UTCTime)

Timestamp of last printed message

data SyncLog event

Logging datatype for information that occurred since the previous SyncLog.

Constructors

SyncLog

Fields

_syncLogStats :: LastSyncStats

Stats since the last syncing log message

_syncLogTracer :: MarconiTrace IO

The pretty-printing tracer

_syncLogTimeSinceLastMsg :: Maybe NominalDiffTime

Time since last syncing log message.

withSyncStats :: MarconiTrace IO -> indexer event -> WithSyncStats indexer event

A smart constructor for WithSyncStats