Marconi.ChainIndex.Types

Description

This module provides several type aliases and utility functions to deal with them.

Config for retrying

data RetryConfig

Config type for node retries

Constructors

RetryConfig

Fields

baseTimeBeforeNextRetry :: !Word64

Initial time before next retry (in seconds)

maybeMaxWaitTime :: !(Maybe Word64)

Max time before stopping retries (in seconds)

Application tracer

type MarconiTrace m = Trace m (Doc ())

Alias for a Doc tracer, it is the Trace used throughout the Marconi application

Addresses alias used to query marconi

type TargetAddresses = NESet (Address ShelleyAddr)

Type represents non empty set of Bech32 Shelley compatible addresses

Utxo Indexer Configuration, containing targetAddresses and showReferenceScript flag

data UtxoIndexerConfig

Constructors

UtxoIndexerConfig

Fields

ucTargetAddresses :: Maybe TargetAddresses

List of address utxo indexer to follow

ucEnableUtxoTxOutRef :: Bool

enable utxo indexer to store txOut refScript

Aliases for the current Cardano era

type CurrentEra = BabbageEra

An alias for the current era, to ease the transition from one era to the next one

pattern AsCurrentEra :: AsType CurrentEra

pattern CurrentEra :: CardanoEra CurrentEra

type TxOut = TxOut CtxTx CurrentEra

A Cardano TxOut of the current Era

Aliases to ease concept mapping between plutus types and cardano types

type TxOutRef = TxIn

A reference to a transaction output. This is a pair of a transaction reference, and an index indicating which of the outputs of that transaction we are referring to.

txOutRef :: TxId -> TxIx -> TxIn

Reexport from cardano-api-extended

data BlockEvent

A block, along with some contextual information

Constructors

BlockEvent

Fields

blockInMode :: BlockInMode CardanoMode

epochNo :: EpochNo

blockTime :: POSIXTime

Database file names

utxoDbName :: FilePath

addressDatumDbName :: FilePath

scriptTxDbName :: FilePath

epochStateDbName :: FilePath

mintBurnDbName :: FilePath

newtype SecurityParam

Constructors

SecurityParam Word64

data IndexingDepth

Constructors

MinIndexingDepth !Word64

MaxIndexingDepth

newtype TxIndexInBlock

Constructors

TxIndexInBlock Word64

newtype ShouldFailIfResync

Constructors

ShouldFailIfResync Bool

Configuration for index runners

data RunIndexerConfig

Common configuration required to run indexers

Constructors

RunIndexerConfig (MarconiTrace IO) RetryConfig SecurityParam NetworkId ChainPoint FilePath

runIndexerConfigTrace :: Lens' RunIndexerConfig (MarconiTrace IO)

runIndexerConfigRetryConfig :: Lens' RunIndexerConfig RetryConfig

runIndexerConfigSecurityParam :: Lens' RunIndexerConfig SecurityParam

runIndexerConfigNetworkId :: Lens' RunIndexerConfig NetworkId

runIndexerConfigChainPoint :: Lens' RunIndexerConfig ChainPoint

runIndexerConfigSocketPath :: Lens' RunIndexerConfig FilePath