Marconi.ChainIndex.Utils

Description

This module contains a bunch of utility functions for working in `marconi-chain-index`.

Documentation

isBlockRollbackable :: SecurityParam -> BlockNo -> BlockNo -> Bool

getBlockNoFromChainTip :: ChainTip -> BlockNo

queryCurrentNodeBlockNo

Arguments

:: NetworkId

-> FilePath
Node socket file path

-> ExceptT (IndexerError err) IO BlockNo

Query the current tip from the local node.

querySecurityParam

Arguments

:: NetworkId

-> FilePath
Node socket file path

-> ExceptT (IndexerError err) IO SecurityParam

Query security param from the local node. It queries the current era first, and uses that to query the security parameter.

querySecurityParamEra

Arguments

:: forall era err. ShelleyBasedEra era

-> NetworkId

-> FilePath
Node socket file path

-> ExceptT (IndexerError err) IO SecurityParam

Query security param from the local node given a Shelley based era.

toException :: Exception err => ExceptT err IO a -> IO a

chainPointOrGenesis :: [ChainPoint] -> ChainPoint

Return the first element of the list of chain points. If the list is empty, return the genesis point.

addressesToPredicate :: Maybe TargetAddresses -> Maybe (Address ShelleyAddr -> Bool)

Convert a list of target addresses to a predicate function

chainPointToSlotNo :: ChainPoint -> Maybe SlotNo

chainPointToHash :: ChainPoint -> Maybe (Hash BlockHeader)