Marconi.ChainIndex.Indexers

Documentation

getDatums :: BlockInMode CardanoMode -> [(SlotNo, (Hash ScriptData, ScriptData))]

scriptDataFromCardanoTxBody :: TxBody era -> Map (Hash ScriptData) ScriptData

data Coordinator

The way we synchronise channel consumption is by waiting on a QSemN for each of the spawn indexers to finish processing the current event.

The channel is used to transmit the next event to the listening indexers. Note that even if the channel is unbound it will actually only ever hold one event because it will be blocked until the processing of the event finishes on all indexers.

The indexer count is where we save the number of running indexers so we know for how many we are waiting.

Constructors

Coordinator

Fields

_channel :: !(TChan (ChainSyncEvent (BlockInMode CardanoMode)))

_barrier :: !QSemN

_indexerCount :: !Int

initialCoordinator :: Int -> IO Coordinator

type Worker = SecurityParam -> Coordinator -> FilePath -> IO [StorablePoint ScriptTxHandle]

datumWorker :: Worker

utxoWorker_

Arguments

:: (UtxoIndexer -> IO ())
callback function used in the queryApi thread, needs to be non-blocking

-> Depth

-> Maybe TargetAddresses
Target addresses to filter for

-> Coordinator

-> TChan (ChainSyncEvent (BlockInMode CardanoMode))

-> FilePath

-> IO (IO (), MVar UtxoIndexer)

utxoWorker

Arguments

:: (UtxoIndexer -> IO ())
callback function used in the queryApi thread

-> Maybe TargetAddresses
Target addresses to filter for

-> Worker

addressDatumWorker :: (StorableEvent AddressDatumHandle -> IO [()]) -> Maybe TargetAddresses -> Worker

addressDatumWorker_

Arguments

:: (StorableEvent AddressDatumHandle -> IO [()])

-> Maybe TargetAddresses
Target addresses to filter for

-> AddressDatumDepth

-> Coordinator

-> TChan (ChainSyncEvent (BlockInMode CardanoMode))

-> FilePath

-> IO (IO (), MVar AddressDatumIndex)

ScriptTx indexer

scriptTxWorker_ :: (StorableEvent ScriptTxHandle -> IO [()]) -> Depth -> Coordinator -> TChan (ChainSyncEvent (BlockInMode CardanoMode)) -> FilePath -> IO (IO (), MVar ScriptTxIndexer)

scriptTxWorker :: (StorableEvent ScriptTxHandle -> IO [()]) -> Worker

Epoch state indexer

epochStateWorker_ :: FilePath -> ((State EpochStateHandle, StorableEvent EpochStateHandle) -> IO ()) -> SecurityParam -> Coordinator -> TChan (ChainSyncEvent (BlockInMode CardanoMode)) -> FilePath -> IO (IO b, MVar EpochStateIndex)

epochStateWorker :: FilePath -> ((State EpochStateHandle, StorableEvent EpochStateHandle) -> IO ()) -> Worker

Mint/burn indexer

mintBurnWorker_ :: SecurityParam -> (TxMintEvent -> IO ()) -> Coordinator -> TChan (ChainSyncEvent (BlockInMode CardanoMode)) -> FilePath -> IO (IO b, MVar MintBurnIndexer)

mintBurnWorker :: (TxMintEvent -> IO ()) -> Worker

initializeIndexers :: SecurityParam -> [(Worker, FilePath)] -> IO ([ChainPoint], Coordinator)

mkIndexerStream :: Coordinator -> Stream (Of (ChainSyncEvent (BlockInMode CardanoMode))) IO r -> IO ()

runIndexers :: FilePath -> NetworkId -> ChainPoint -> Text -> [(Worker, Maybe FilePath)] -> IO ()