Marconi.ChainIndex.Indexers.EpochState

Events

data EpochState

Base event used to store the ExtLedgerState

Constructors

EpochState ExtLedgerState BlockNo

type ExtLedgerState = ExtLedgerState (HardForkBlock (CardanoEras StandardCrypto))

extLedgerState :: EpochState -> ExtLedgerState

blockNo :: EpochState -> BlockNo

data EpochNonce

Event for Nonce storage

Constructors

EpochNonce !EpochNo !Nonce !BlockNo

nonceEpochNo :: Lens' EpochNonce EpochNo

nonceNonce :: Lens' EpochNonce Nonce

nonceBlockNo :: Lens' EpochNonce BlockNo

data EpochSDD

Event for SDD storage

Constructors

EpochSDD !EpochNo !PoolId !Lovelace !BlockNo

sddEpochNo :: Lens' EpochSDD EpochNo

sddPoolId :: Lens' EpochSDD PoolId

sddLovelace :: Lens' EpochSDD Lovelace

sddBlockNo :: Lens' EpochSDD BlockNo

data EpochMetadata

Metadata used to cerate EpochStateIndexer filenames

Constructors

EpochMetadata

Fields

metadataBlockNo :: Maybe BlockNo

metadataChainpoint :: ChainPoint

Indexer and worker

data EpochStateIndexer event

The main type of the epoch state indexer, it contains both the indexers it operates and its configuration

newtype NodeConfig

Constructors

NodeConfig

Fields

nodeConfig :: FilePath

node config path

data EpochStateWorkerConfig

Constructors

EpochStateWorkerConfig

Fields

indexerConfig :: NodeConfig

epochSnapshotInterval :: Word

type StandardEpochStateIndexer = EpochStateIndexer (WithDistance (BlockInMode CardanoMode))

mkEpochStateIndexer :: (MonadIO n, MonadError IndexerError n) => StandardWorkerConfig IO a b -> NodeConfig -> FilePath -> n (EpochStateIndexer (WithDistance (Maybe ExtLedgerState, BlockInMode CardanoMode)))

mkEpochStateWorker

Arguments

:: forall m input. (MonadIO m, MonadError IndexerError m)

=> StandardWorkerConfig IO input (BlockInMode CardanoMode)
General configuration of the indexer (mostly for logging purpose)

-> EpochStateWorkerConfig

-> FilePath

-> m (WorkerIndexer IO (WithDistance input) (WithDistance (Maybe ExtLedgerState, BlockInMode CardanoMode)) (WithTrace IO EpochStateIndexer))

Queries

data ActiveSDDByEpochNoQuery

data NonceByEpochNoQuery